move Haskell to own dir
