+++ packages/quicktype-core/src/language/Haskell/constants.ts
@@ -1,0 +1,54 @@
+export const forbiddenNames = [
+    // reserved keywords
+    "as",
+    "case",
+    "class",
+    "data",
+    "default",
+    "deriving",
+    "do",
+    "else",
+    "family",
+    "forall",
+    "foreign",
+    "hiding",
+    "if",
+    "import",
+    "in",
+    "infix",
+    "infixl",
+    "infixr",
+    "instance",
+    "let",
+    "of",
+    "mdo",
+    "module",
+    "newtype",
+    "proc",
+    "qualified",
+    "rec",
+    "then",
+    "type",
+    "where",
+    // in Prelude keywords ...
+    "id",
+    "Array",
+    "HashMap",
+    "Map",
+    "Maybe",
+    "Bool",
+    "Int",
+    "True",
+    "False",
+    "Enum",
+    // Aeson types
+    "encode",
+    "decode",
+    "text",
+    "Text",
+    "Value",
+    "Object",
+    "Result",
+    "Series",
+    "Error"
+] as const;
