+++ packages/quicktype-core/src/language/Haskell/index.ts
@@ -1,0 +1,2 @@
+export { HaskellTargetLanguage, haskellOptions } from "./language";
+export { HaskellRenderer } from "./HaskellRenderer";
